requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
openai==1.55.0
httpx==0.27.0
aiohttp==3.10.0
python-dotenv==1.0.0
tiktoken==0.8.0
